requirements.txt
streamlit
streamlit-authenticator
requests
pandas
boto3
python-dotenv